semantic-service/requirements-dev.txt
flask>=3.0.0
sentence-transformers>=2.7.0
numpy>=1.24.0
transformers>=4.40.0
scikit-learn>=1.3.0
gunicorn>=21.2.0

# PyTorch (regular version for Mac development)
torch>=2.1.0,<3.0.0

# Evaluation framework dependencies
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
requests>=2.28.0

# Hybrid approach - OpenAI API
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Emoji handling
emoji>=2.8.0